+++ services/gateway/sources/version3/mappings/dposVote.js
